Add new pointer types for a service
