+++ remote-trait-object/src/service/serde_support.rs
@@ -15,6 +15,7 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 use super::*;
+use super::export_import::*;
 use serde::de::{Deserialize, Deserializer};
 use serde::ser::{Serialize, Serializer};
